pyproject.toml
[project]
name = "weekly-briefing-agent"
version = "0.1.0"
description = "Construction industry weekly briefing PPT auto-generator"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.99",
    "openai-agents>=0.3",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "playwright>=1.48",
    "pdfplumber>=0.11",
    "pypdf>=5.0",
    "pdf2image>=1.17",
    "pytesseract>=0.3.13",
    "python-pptx>=1.0",
    "matplotlib>=3.9",
    "pandas>=2.2",
    "kiwipiepy>=0.21",
    "tenacity>=9.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.32",
    "respx>=0.21",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "ASYNC"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "agents.*",
    "bs4.*",
    "openai.*",
    "pdfplumber.*",
    "pdf2image.*",
    "pptx.*",
    "pytesseract.*",
    "kiwipiepy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
